Fixed warnings

+++ testbed/tests/CuTest/CuTest.c
@@ -13,13 +13,13 @@
 
 char* CuStrAlloc(int size)
 {
-	char* newStr = (char*) malloc( sizeof(char) * (size) );
+	char* newStr = (char*) malloc( sizeof(char) * (size_t)size );
 	return newStr;
 }
 
 char* CuStrCopy(const char* old)
 {
-	int len = strlen(old);
+	int len = (int)strlen(old);
 	char* newStr = CuStrAlloc(len + 1);
 	strcpy(newStr, old);
 	return newStr;
@@ -33,7 +33,7 @@ void CuStringInit(CuString* str)
 {
 	str->length = 0;
 	str->size = STRING_MAX;
-	str->buffer = (char*) malloc(sizeof(char) * str->size);
+	str->buffer = (char*) malloc(sizeof(char) * (size_t)str->size);
 	str->buffer[0] = '\0';
 }
 
@@ -42,7 +42,7 @@ CuString* CuStringNew(void)
 	CuString* str = (CuString*) malloc(sizeof(CuString));
 	str->length = 0;
 	str->size = STRING_MAX;
-	str->buffer = (char*) malloc(sizeof(char) * str->size);
+	str->buffer = (char*) malloc(sizeof(char) * (size_t)str->size);
 	str->buffer[0] = '\0';
 	return str;
 }
@@ -56,7 +56,7 @@ void CuStringDelete(CuString *str)
 
 void CuStringResize(CuString* str, int newSize)
 {
-	str->buffer = (char*) realloc(str->buffer, sizeof(char) * newSize);
+	str->buffer = (char*) realloc(str->buffer, sizeof(char) * (size_t)newSize);
 	str->size = newSize;
 }
 
@@ -68,7 +68,7 @@ void CuStringAppend(CuString* str, const char* text)
 		text = "NULL";
 	}
 
-	length = strlen(text);
+	length = (int)strlen(text);
 	if (str->length + length + 1 >= str->size)
 		CuStringResize(str, str->length + length + 1 + STRING_INC);
 	str->length += length;
@@ -95,7 +95,7 @@ void CuStringAppendFormat(CuString* str, const char* format, ...)
 
 void CuStringInsert(CuString* str, const char* text, int pos)
 {
-	int length = strlen(text);
+	int length = (int)strlen(text);
 	if (pos > str->length)
 		pos = str->length;
 	if (str->length + length + 1 >= str->size)
